+++ src/main/java/ch/technokrat/gecko/geckocircuits/control/ControlMUX.java
@@ -103,15 +103,15 @@ public final class ControlMUX extends ControlBlock implements VariableTerminalNu
         xKlickMin = (int) (dpix * (posX - WIDTH));
         xKlickMax = (int) (dpix * (posX + WIDTH));
         yKlickMin = (int) (dpix * (posY - WIDTH));
-        yKlickMax = (int) (dpix * (posY + XIN.size()));
+        yKlickMax = dpix * (posY + XIN.size());
         graphics.setColor(getBackgroundColor());
 
         graphics.fillRect((int) (dpix * (posX - WIDTH)), (int) (dpix * (posY - WIDTH)),
-                (int) (dpix * (2 * WIDTH)), (int) (dpix * XIN.size()));
+                (int) (dpix * (2 * WIDTH)), dpix * XIN.size());
 
         graphics.setColor(origColor);
         graphics.drawRect((int) (dpix * (posX - WIDTH)), (int) (dpix * (posY - WIDTH)),
-                (int) (dpix * (2 * WIDTH)), (int) (dpix * XIN.size()));
+                (int) (dpix * (2 * WIDTH)), dpix * XIN.size());
         // Pfeil-Symbol:
         int d1 = 10, d2 = 4, dpfx = 8, dpfy = 3;
         double pf = 1.4;  // Pfeilspitzen-X-Abstand
